# remove redundant isNaN check and add fast path for empty options