--- a/cookie.js
+++ b/cookie.js
@@ -121,7 +121,7 @@ function parse (str, opt) {
  */
 
 function serialize (name, val, opt) {
-  if (name?.length && !fieldContentRegExp.test(name)) {
+  if (!fieldContentRegExp.test(name)) {
     throw new TypeError('argument name is invalid')
   }