--- a/cookie.js
+++ b/cookie.js
@@ -29,14 +29,6 @@
 'use strict'
 
 /**
- * Module exports.
- * @public
- */
-
-exports.parse = parse
-exports.serialize = serialize
-
-/**
  * RegExp to match field-content in RFC 7230 sec 3.2
  *
  * field-content = field-vchar [ 1*( SP / HTAB ) field-vchar ]